Move ScryptoValue to scrypto standard lib
