+++ radix-engine/src/engine/process.rs
@@ -10,6 +10,7 @@ use scrypto::rust::format;
 use scrypto::rust::string::String;
 use scrypto::rust::vec;
 use scrypto::rust::vec::Vec;
+use scrypto::values::*;
 use wasmi::*;
 
 use crate::engine::process::LazyMapState::{Committed, Uncommitted};
@@ -64,7 +65,7 @@ pub struct Invocation {
     package_address: PackageAddress,
     export_name: String,
     function: String,
-    args: Vec<ValidatedData>,
+    args: Vec<ScryptoValue>,
 }
 
 /// Qualitative states for a WASM process
@@ -301,10 +302,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
     }
 
     // (Transaction ONLY) Returns resource back to worktop.
-    pub fn return_to_worktop(
-        &mut self,
-        bucket_id: BucketId,
-    ) -> Result<ValidatedData, RuntimeError> {
+    pub fn return_to_worktop(&mut self, bucket_id: BucketId) -> Result<ScryptoValue, RuntimeError> {
         re_debug!(
             self,
             "(Transaction) Returning to worktop: bucket_id = {}",
@@ -318,18 +316,18 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         self.worktop
             .put(bucket)
             .map_err(RuntimeError::WorktopError)?;
-        Ok(ValidatedData::from_value(&()))
+        Ok(ScryptoValue::from_value(&()))
     }
 
     // (Transaction ONLY) Assert worktop contains at least this amount.
     pub fn assert_worktop_contains(
         &mut self,
         resource_address: ResourceAddress,
-    ) -> Result<ValidatedData, RuntimeError> {
+    ) -> Result<ScryptoValue, RuntimeError> {
         if self.worktop.total_amount(resource_address).is_zero() {
             Err(RuntimeError::AssertionFailed)
         } else {
-            Ok(ValidatedData::from_value(&()))
+            Ok(ScryptoValue::from_value(&()))
         }
     }
 
@@ -338,11 +336,11 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         &mut self,
         amount: Decimal,
         resource_address: ResourceAddress,
-    ) -> Result<ValidatedData, RuntimeError> {
+    ) -> Result<ScryptoValue, RuntimeError> {
         if self.worktop.total_amount(resource_address) < amount {
             Err(RuntimeError::AssertionFailed)
         } else {
-            Ok(ValidatedData::from_value(&()))
+            Ok(ScryptoValue::from_value(&()))
         }
     }
 
@@ -351,7 +349,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         &mut self,
         ids: &BTreeSet<NonFungibleId>,
         resource_address: ResourceAddress,
-    ) -> Result<ValidatedData, RuntimeError> {
+    ) -> Result<ScryptoValue, RuntimeError> {
         if !self
             .worktop
             .total_ids(resource_address)
@@ -360,7 +358,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         {
             Err(RuntimeError::AssertionFailed)
         } else {
-            Ok(ValidatedData::from_value(&()))
+            Ok(ScryptoValue::from_value(&()))
         }
     }
 
@@ -623,7 +621,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         &mut self,
         component_address: ComponentAddress,
         method: &str,
-    ) -> Result<ValidatedData, RuntimeError> {
+    ) -> Result<ScryptoValue, RuntimeError> {
         re_debug!(
             self,
             "(Transaction) Calling method with all resources started"
@@ -660,7 +658,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         let invocation = self.prepare_call_method(
             component_address,
             method,
-            vec![ValidatedData::from_slice(&scrypto_encode(&to_deposit)).unwrap()],
+            vec![ScryptoValue::from_slice(&scrypto_encode(&to_deposit)).unwrap()],
         )?;
         let result = self.call(invocation);
 
@@ -694,7 +692,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
     }
 
     /// Runs the given export within this process.
-    pub fn run(&mut self, invocation: Invocation) -> Result<ValidatedData, RuntimeError> {
+    pub fn run(&mut self, invocation: Invocation) -> Result<ScryptoValue, RuntimeError> {
         #[cfg(not(feature = "alloc"))]
         let now = std::time::Instant::now();
         re_info!(
@@ -814,7 +812,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         package_address: PackageAddress,
         blueprint_name: &str,
         function: &str,
-        args: Vec<ValidatedData>,
+        args: Vec<ScryptoValue>,
     ) -> Result<Invocation, RuntimeError> {
         Ok(Invocation {
             actor: Actor::Blueprint(blueprint_name.to_owned()),
@@ -830,7 +828,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         &mut self,
         component_address: ComponentAddress,
         method: &str,
-        args: Vec<ValidatedData>,
+        args: Vec<ScryptoValue>,
     ) -> Result<Invocation, RuntimeError> {
         let component = self
             .track
@@ -838,7 +836,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
             .ok_or(RuntimeError::ComponentNotFound(component_address))?
             .clone();
         let mut args_with_self =
-            vec![ValidatedData::from_slice(&scrypto_encode(&component_address)).unwrap()];
+            vec![ScryptoValue::from_slice(&scrypto_encode(&component_address)).unwrap()];
         args_with_self.extend(args);
 
         Ok(Invocation {
@@ -866,7 +864,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
     }
 
     /// Calls a function/method.
-    pub fn call(&mut self, invocation: Invocation) -> Result<ValidatedData, RuntimeError> {
+    pub fn call(&mut self, invocation: Invocation) -> Result<ScryptoValue, RuntimeError> {
         // figure out what buckets and proofs to move from this process
         let mut moving_buckets = HashMap::new();
         let mut moving_proofs = HashMap::new();
@@ -908,8 +906,8 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         package_address: PackageAddress,
         blueprint_name: &str,
         function: &str,
-        args: Vec<ValidatedData>,
-    ) -> Result<ValidatedData, RuntimeError> {
+        args: Vec<ScryptoValue>,
+    ) -> Result<ScryptoValue, RuntimeError> {
         re_debug!(self, "Call function started");
         let invocation =
             self.prepare_call_function(package_address, blueprint_name, function, args)?;
@@ -923,8 +921,8 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         &mut self,
         component_address: ComponentAddress,
         method: &str,
-        args: Vec<ValidatedData>,
-    ) -> Result<ValidatedData, RuntimeError> {
+        args: Vec<ScryptoValue>,
+    ) -> Result<ScryptoValue, RuntimeError> {
         re_debug!(self, "Call method started");
         let invocation = self.prepare_call_method(component_address, method, args)?;
         let result = self.call(invocation);
@@ -937,7 +935,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         &mut self,
         package_address: PackageAddress,
         blueprint_name: &str,
-    ) -> Result<ValidatedData, RuntimeError> {
+    ) -> Result<ScryptoValue, RuntimeError> {
         re_debug!(self, "Call abi started");
         let invocation = self.prepare_call_abi(package_address, blueprint_name)?;
         let result = self.call(invocation);
@@ -1053,7 +1051,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         Ok(())
     }
 
-    fn process_call_data(&mut self, validated: &ValidatedData) -> Result<(), RuntimeError> {
+    fn process_call_data(&mut self, validated: &ScryptoValue) -> Result<(), RuntimeError> {
         if !validated.lazy_map_ids.is_empty() {
             return Err(RuntimeError::LazyMapNotAllowed);
         }
@@ -1063,7 +1061,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         Ok(())
     }
 
-    fn process_return_data(&mut self, validated: &ValidatedData) -> Result<(), RuntimeError> {
+    fn process_return_data(&mut self, validated: &ScryptoValue) -> Result<(), RuntimeError> {
         if !validated.lazy_map_ids.is_empty() {
             return Err(RuntimeError::LazyMapNotAllowed);
         }
@@ -1076,7 +1074,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
     /// Process and parse entry data from any component object (components and maps)
     fn process_entry_data(data: &[u8]) -> Result<ComponentObjectRefs, RuntimeError> {
         let validated =
-            ValidatedData::from_slice(data).map_err(RuntimeError::DataValidationError)?;
+            ScryptoValue::from_slice(data).map_err(RuntimeError::ParseScryptoValueError)?;
         if !validated.bucket_ids.is_empty() {
             return Err(RuntimeError::BucketNotAllowed);
         }
@@ -1108,9 +1106,9 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         })
     }
 
-    fn process_non_fungible_data(&mut self, data: &[u8]) -> Result<ValidatedData, RuntimeError> {
+    fn process_non_fungible_data(&mut self, data: &[u8]) -> Result<ScryptoValue, RuntimeError> {
         let validated =
-            ValidatedData::from_slice(data).map_err(RuntimeError::DataValidationError)?;
+            ScryptoValue::from_slice(data).map_err(RuntimeError::ParseScryptoValueError)?;
         if !validated.bucket_ids.is_empty() {
             return Err(RuntimeError::BucketNotAllowed);
         }
@@ -1222,7 +1220,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
         Err(RuntimeError::MemoryAllocError)
     }
 
-    fn read_return_value(&mut self, ptr: u32) -> Result<ValidatedData, RuntimeError> {
+    fn read_return_value(&mut self, ptr: u32) -> Result<ScryptoValue, RuntimeError> {
         let wasm_process = self.wasm_process_state.as_ref().unwrap();
         // read length
         let len: u32 = wasm_process
@@ -1243,7 +1241,7 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
             return Err(RuntimeError::MemoryAccessError);
         }
 
-        ValidatedData::from_slice(&buffer[range]).map_err(RuntimeError::DataValidationError)
+        ScryptoValue::from_slice(&buffer[range]).map_err(RuntimeError::ParseScryptoValueError)
     }
 
     /// Handles a system call.
@@ -1306,8 +1304,9 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
     ) -> Result<CallFunctionOutput, RuntimeError> {
         let mut validated_args = Vec::new();
         for arg in input.args {
-            validated_args
-                .push(ValidatedData::from_slice(&arg).map_err(RuntimeError::DataValidationError)?);
+            validated_args.push(
+                ScryptoValue::from_slice(&arg).map_err(RuntimeError::ParseScryptoValueError)?,
+            );
         }
 
         re_debug!(
@@ -1337,8 +1336,9 @@ impl<'r, 'l, L: SubstateStore> Process<'r, 'l, L> {
     ) -> Result<CallMethodOutput, RuntimeError> {
         let mut validated_args = Vec::new();
         for arg in input.args {
-            validated_args
-                .push(ValidatedData::from_slice(&arg).map_err(RuntimeError::DataValidationError)?);
+            validated_args.push(
+                ScryptoValue::from_slice(&arg).map_err(RuntimeError::ParseScryptoValueError)?,
+            );
         }
 
         re_debug!(
