+++ radix-engine/src/errors.rs
@@ -3,7 +3,7 @@ use scrypto::crypto::*;
 use scrypto::engine::types::*;
 use scrypto::rust::fmt;
 use scrypto::rust::string::String;
-use scrypto::types::*;
+use scrypto::values::*;
 use wasmi::*;
 
 use crate::engine::*;
@@ -66,17 +66,10 @@ pub enum WasmValidationError {
     InvalidPackageInit,
 }
 
-/// Represents an error when parsing a value from a byte array.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DataValidationError {
-    DecodeError(DecodeError),
-    ScryptoTypeValidationError(ScryptoTypeValidationError),
-}
-
 /// Represents an error when validating a transaction.
 #[derive(Debug, PartialEq, Eq)]
 pub enum TransactionValidationError {
-    DataValidationError(DataValidationError),
+    ParseScryptoValueError(ParseScryptoValueError),
     IdValidatorError(IdValidatorError),
     VaultNotAllowed(VaultId),
     LazyMapNotAllowed(LazyMapId),
@@ -94,7 +87,7 @@ pub enum RuntimeError {
     WasmValidationError(WasmValidationError),
 
     /// The data is not a valid SBOR value.
-    DataValidationError(DataValidationError),
+    ParseScryptoValueError(ParseScryptoValueError),
 
     /// Not a valid ABI.
     AbiValidationError(DecodeError),
