+++ radix-engine/src/model/receipt.rs
@@ -4,6 +4,7 @@ use scrypto::rust::fmt;
 use scrypto::rust::string::String;
 use scrypto::rust::string::ToString;
 use scrypto::rust::vec::Vec;
+use scrypto::values::*;
 
 use crate::engine::CommitReceipt;
 use crate::errors::*;
@@ -14,7 +15,7 @@ pub struct Receipt {
     pub commit_receipt: Option<CommitReceipt>,
     pub transaction: ValidatedTransaction,
     pub result: Result<(), RuntimeError>,
-    pub outputs: Vec<ValidatedData>,
+    pub outputs: Vec<ScryptoValue>,
     pub logs: Vec<(Level, String)>,
     pub new_package_addresses: Vec<PackageAddress>,
     pub new_component_addresses: Vec<ComponentAddress>,
