+++ radix-engine/src/transaction/executor.rs
@@ -5,6 +5,7 @@ use scrypto::engine::types::*;
 use scrypto::resource::ProofRule;
 use scrypto::rust::vec;
 use scrypto::rust::vec::Vec;
+use scrypto::values::*;
 use scrypto::{abi, this};
 
 use crate::engine::*;
@@ -190,7 +191,7 @@ impl<'l, L: SubstateStore> TransactionExecutor<'l, L> {
                 ValidatedInstruction::TakeFromWorktop { resource_address } => proc
                     .take_all_from_worktop(*resource_address)
                     .map(|bucket_id| {
-                        ValidatedData::from_value(&scrypto::resource::Bucket(bucket_id))
+                        ScryptoValue::from_value(&scrypto::resource::Bucket(bucket_id))
                     }),
                 ValidatedInstruction::TakeFromWorktopByAmount {
                     amount,
@@ -198,7 +199,7 @@ impl<'l, L: SubstateStore> TransactionExecutor<'l, L> {
                 } => proc
                     .take_from_worktop(*amount, *resource_address)
                     .map(|bucket_id| {
-                        ValidatedData::from_value(&scrypto::resource::Bucket(bucket_id))
+                        ScryptoValue::from_value(&scrypto::resource::Bucket(bucket_id))
                     }),
                 ValidatedInstruction::TakeFromWorktopByIds {
                     ids,
@@ -206,7 +207,7 @@ impl<'l, L: SubstateStore> TransactionExecutor<'l, L> {
                 } => proc
                     .take_non_fungibles_from_worktop(ids, *resource_address)
                     .map(|bucket_id| {
-                        ValidatedData::from_value(&scrypto::resource::Bucket(bucket_id))
+                        ScryptoValue::from_value(&scrypto::resource::Bucket(bucket_id))
                     }),
                 ValidatedInstruction::ReturnToWorktop { bucket_id } => {
                     proc.return_to_worktop(*bucket_id)
@@ -224,37 +225,37 @@ impl<'l, L: SubstateStore> TransactionExecutor<'l, L> {
                 } => proc.assert_worktop_contains_by_ids(&ids, *resource_address),
                 ValidatedInstruction::PopFromAuthZone {} => proc
                     .pop_from_auth_zone()
-                    .map(|proof_id| ValidatedData::from_value(&scrypto::resource::Proof(proof_id))),
+                    .map(|proof_id| ScryptoValue::from_value(&scrypto::resource::Proof(proof_id))),
                 ValidatedInstruction::ClearAuthZone => proc
                     .drop_all_auth_zone_proofs()
-                    .map(|_| ValidatedData::from_value(&())),
+                    .map(|_| ScryptoValue::from_value(&())),
                 ValidatedInstruction::PushToAuthZone { proof_id } => proc
                     .push_to_auth_zone(*proof_id)
-                    .map(|_| ValidatedData::from_value(&())),
+                    .map(|_| ScryptoValue::from_value(&())),
                 ValidatedInstruction::CreateProofFromAuthZone { resource_address } => proc
                     .create_auth_zone_proof(*resource_address)
-                    .map(|proof_id| ValidatedData::from_value(&scrypto::resource::Proof(proof_id))),
+                    .map(|proof_id| ScryptoValue::from_value(&scrypto::resource::Proof(proof_id))),
                 ValidatedInstruction::CreateProofFromAuthZoneByAmount {
                     amount,
                     resource_address,
                 } => proc
                     .create_auth_zone_proof_by_amount(*amount, *resource_address)
-                    .map(|proof_id| ValidatedData::from_value(&scrypto::resource::Proof(proof_id))),
+                    .map(|proof_id| ScryptoValue::from_value(&scrypto::resource::Proof(proof_id))),
                 ValidatedInstruction::CreateProofFromAuthZoneByIds {
                     ids,
                     resource_address,
                 } => proc
                     .create_auth_zone_proof_by_ids(ids, *resource_address)
-                    .map(|proof_id| ValidatedData::from_value(&scrypto::resource::Proof(proof_id))),
+                    .map(|proof_id| ScryptoValue::from_value(&scrypto::resource::Proof(proof_id))),
                 ValidatedInstruction::CreateProofFromBucket { bucket_id } => proc
                     .create_bucket_proof(*bucket_id)
-                    .map(|proof_id| ValidatedData::from_value(&scrypto::resource::Proof(proof_id))),
+                    .map(|proof_id| ScryptoValue::from_value(&scrypto::resource::Proof(proof_id))),
                 ValidatedInstruction::CloneProof { proof_id } => proc
                     .clone_proof(*proof_id)
-                    .map(|proof_id| ValidatedData::from_value(&scrypto::resource::Proof(proof_id))),
+                    .map(|proof_id| ScryptoValue::from_value(&scrypto::resource::Proof(proof_id))),
                 ValidatedInstruction::DropProof { proof_id } => proc
                     .drop_proof(*proof_id)
-                    .map(|_| ValidatedData::from_value(&())),
+                    .map(|_| ScryptoValue::from_value(&())),
                 ValidatedInstruction::CallFunction {
                     package_address,
                     blueprint_name,
@@ -272,7 +273,7 @@ impl<'l, L: SubstateStore> TransactionExecutor<'l, L> {
                 } => proc.call_method_with_all_resources(*component_address, &method),
                 ValidatedInstruction::PublishPackage { code } => proc
                     .publish_package(code.clone())
-                    .map(|package_address| ValidatedData::from_value(&package_address)),
+                    .map(|package_address| ScryptoValue::from_value(&package_address)),
             };
             match result {
                 Ok(data) => {
