+++ radix-engine/tests/package.rs
@@ -160,7 +160,7 @@ fn zero_return_len_should_cause_data_validation_error() {
 
     // Assert
     let error = receipt.result.expect_err("Should be an error.");
-    if !matches!(error, RuntimeError::DataValidationError(_)) {
+    if !matches!(error, RuntimeError::ParseScryptoValueError(_)) {
         panic!("{} should be data validation error", error);
     }
 }
