+++ scrypto/src/types.rs
@@ -1,16 +1,5 @@
-use sbor::{any::*, *};
-
-use crate::component::*;
-use crate::crypto::*;
-use crate::engine::types::BucketId;
-use crate::engine::types::ProofId;
-use crate::math::*;
-use crate::resource::*;
 use crate::rust::borrow::ToOwned;
-use crate::rust::collections::HashMap;
-use crate::rust::format;
 use crate::rust::string::String;
-use crate::rust::vec::Vec;
 
 /// A macro to help create a Scrypto-specific type.
 macro_rules! scrypto_type {
@@ -126,184 +115,5 @@ impl ScryptoType {
             .next()
             .unwrap()
             .to_owned()
-    }
-}
-
-/// A validator the check a Scrypto-specific value.
-pub struct ScryptoTypeValidator {
-    pub buckets: Vec<Bucket>,
-    pub proofs: Vec<Proof>,
-    pub vaults: Vec<Vault>,
-    pub lazy_maps: Vec<LazyMap<(), ()>>,
-}
-
-/// Represents an error when validating a Scrypto-specific value.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ScryptoTypeValidationError {
-    DecodeError(DecodeError),
-    InvalidTypeId(u8),
-    InvalidDecimal(ParseDecimalError),
-    InvalidPackageAddress(ParsePackageAddressError),
-    InvalidComponentAddress(ParseComponentAddressError),
-    InvalidResourceAddress(ParseResourceAddressError),
-    InvalidHash(ParseHashError),
-    InvalidEcdsaPublicKey(ParseEcdsaPublicKeyError),
-    InvalidEcdsaSignature(ParseEcdsaSignatureError),
-    InvalidBucket(ParseBucketError),
-    InvalidProof(ParseProofError),
-    InvalidLazyMap(ParseLazyMapError),
-    InvalidVault(ParseVaultError),
-    InvalidNonFungibleId(ParseNonFungibleIdError),
-    InvalidNonFungibleAddress(ParseNonFungibleAddressError),
-}
-
-impl ScryptoTypeValidator {
-    pub fn new() -> Self {
-        Self {
-            buckets: Vec::new(),
-            proofs: Vec::new(),
-            vaults: Vec::new(),
-            lazy_maps: Vec::new(),
-        }
-    }
-}
-
-impl CustomValueVisitor for ScryptoTypeValidator {
-    type Err = ScryptoTypeValidationError;
-
-    fn visit(&mut self, type_id: u8, data: &[u8]) -> Result<(), Self::Err> {
-        match ScryptoType::from_id(type_id).ok_or(Self::Err::InvalidTypeId(type_id))? {
-            ScryptoType::PackageAddress => {
-                PackageAddress::try_from(data)
-                    .map_err(ScryptoTypeValidationError::InvalidPackageAddress)?;
-            }
-            ScryptoType::ComponentAddress => {
-                ComponentAddress::try_from(data)
-                    .map_err(ScryptoTypeValidationError::InvalidComponentAddress)?;
-            }
-            ScryptoType::LazyMap => {
-                self.lazy_maps.push(
-                    LazyMap::try_from(data).map_err(ScryptoTypeValidationError::InvalidLazyMap)?,
-                );
-            }
-            ScryptoType::Hash => {
-                Hash::try_from(data).map_err(ScryptoTypeValidationError::InvalidHash)?;
-            }
-            ScryptoType::EcdsaPublicKey => {
-                EcdsaPublicKey::try_from(data)
-                    .map_err(ScryptoTypeValidationError::InvalidEcdsaPublicKey)?;
-            }
-            ScryptoType::EcdsaSignature => {
-                EcdsaSignature::try_from(data)
-                    .map_err(ScryptoTypeValidationError::InvalidEcdsaSignature)?;
-            }
-            ScryptoType::Decimal => {
-                Decimal::try_from(data).map_err(ScryptoTypeValidationError::InvalidDecimal)?;
-            }
-            ScryptoType::Bucket => {
-                self.buckets.push(
-                    Bucket::try_from(data).map_err(ScryptoTypeValidationError::InvalidBucket)?,
-                );
-            }
-            ScryptoType::Proof => {
-                self.proofs
-                    .push(Proof::try_from(data).map_err(ScryptoTypeValidationError::InvalidProof)?);
-            }
-            ScryptoType::Vault => {
-                self.vaults
-                    .push(Vault::try_from(data).map_err(ScryptoTypeValidationError::InvalidVault)?);
-            }
-            ScryptoType::NonFungibleId => {
-                NonFungibleId::try_from(data)
-                    .map_err(ScryptoTypeValidationError::InvalidNonFungibleId)?;
-            }
-            ScryptoType::NonFungibleAddress => {
-                NonFungibleAddress::try_from(data)
-                    .map_err(ScryptoTypeValidationError::InvalidNonFungibleAddress)?;
-            }
-            ScryptoType::ResourceAddress => {
-                ResourceAddress::try_from(data)
-                    .map_err(ScryptoTypeValidationError::InvalidResourceAddress)?;
-            }
-        }
-        Ok(())
-    }
-}
-
-/// A formatter that formats a Scrypto type.
-pub struct ScryptoTypeFormatter {}
-
-impl ScryptoTypeFormatter {
-    /// Format a custom value (checked) using the notation introduced by Transaction Manifest.
-    ///
-    /// # Panics
-    /// If the input data or type id is invalid
-    pub fn format(
-        type_id: u8,
-        data: &[u8],
-        bucket_ids: &HashMap<BucketId, String>,
-        proof_ids: &HashMap<ProofId, String>,
-    ) -> String {
-        match ScryptoType::from_id(type_id).unwrap() {
-            ScryptoType::Decimal => format!("Decimal(\"{}\")", Decimal::try_from(data).unwrap()),
-            ScryptoType::PackageAddress => {
-                format!(
-                    "PackageAddress(\"{}\")",
-                    PackageAddress::try_from(data).unwrap()
-                )
-            }
-            ScryptoType::ComponentAddress => {
-                format!(
-                    "ComponentAddress(\"{}\")",
-                    ComponentAddress::try_from(data).unwrap()
-                )
-            }
-            ScryptoType::LazyMap => format!(
-                "LazyMap(\"{}\")",
-                LazyMap::<(), ()>::try_from(data).unwrap()
-            ),
-            ScryptoType::Hash => format!("Hash(\"{}\")", Hash::try_from(data).unwrap()),
-            ScryptoType::EcdsaPublicKey => {
-                format!(
-                    "EcdsaPublicKey(\"{}\")",
-                    EcdsaPublicKey::try_from(data).unwrap()
-                )
-            }
-            ScryptoType::EcdsaSignature => {
-                format!(
-                    "EcdsaSignature(\"{}\")",
-                    EcdsaSignature::try_from(data).unwrap()
-                )
-            }
-            ScryptoType::Bucket => {
-                let bucket = Bucket::try_from(data).unwrap();
-                if let Some(name) = bucket_ids.get(&bucket.0) {
-                    format!("Bucket(\"{}\")", name)
-                } else {
-                    format!("Bucket({}u32)", bucket.0)
-                }
-            }
-            ScryptoType::Proof => {
-                let proof = Proof::try_from(data).unwrap();
-                if let Some(name) = proof_ids.get(&proof.0) {
-                    format!("Proof(\"{}\")", name)
-                } else {
-                    format!("Proof({}u32)", proof.0)
-                }
-            }
-            ScryptoType::Vault => format!("Vault(\"{}\")", Vault::try_from(data).unwrap()),
-            ScryptoType::NonFungibleId => format!(
-                "NonFungibleId(\"{}\")",
-                NonFungibleId::try_from(data).unwrap()
-            ),
-            ScryptoType::NonFungibleAddress => format!(
-                "NonFungibleAddress(\"{}\")",
-                NonFungibleAddress::try_from(data).unwrap()
-            ),
-            ScryptoType::ResourceAddress => format!(
-                "ResourceAddress(\"{}\")",
-                ResourceAddress::try_from(data).unwrap()
-            ),
-        }
     }
 }
